Add helper functions for converting between GridPosition and other types

+++ src/util.rs
@@ -1,5 +1,5 @@
-pub const GRID_SIZE: (i16, i16) = (8, 8);
-pub const GRID_CELL_SIZE: (i16, i16) = (32, 32);
+pub const GRID_SIZE: (u32, u32) = (8, 8);
+pub const GRID_CELL_SIZE: (u32, u32) = (32, 32);
 
 pub const SCREEN_SIZE: (i32, i32) = (
     (GRID_SIZE.0 * GRID_CELL_SIZE.0) as i32,
